Rename GzipHttpClient to HttpClientMediator as it's the only one
